add Delete confirmation alert

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from "react";
-import { Link } from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
+import Swal from "sweetalert2";
 
 // Redux
 import { useDispatch } from "react-redux";
@@ -13,9 +14,21 @@ export default function Product({ product }) {
   //  Confirmar si desea eliminarlo
   const confirmDeleteProduct = (id) => {
     //  preguntar al usuario
-
-    //  pasarlo al dispatch
-    dispatch(deleteProductAction(id));
+    Swal.fire({
+      title: "¿Estas seguro?",
+      text: "Un producto que se elimina no se puede recuperar!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, Eliminar!",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //  pasarlo al dispatch
+        dispatch(deleteProductAction(id));
+      }
+    });
   };
 
   return (
